scripts/zkteco_sync/requirements.txt
pyzk>=0.9
requests>=2.25.0
python-dotenv>=0.19.0
